chore: rename lib/gui/messages => lib/gui/Messages
