+++ src/apps/deskflow-gui/MainWindow.cpp
@@ -19,9 +19,9 @@
 #include "gui/Constants.h"
 #include "gui/Diagnostic.h"
 #include "gui/Logger.h"
+#include "gui/Messages.h"
 #include "gui/core/CoreProcess.h"
 #include "gui/ipc/DaemonIpcClient.h"
-#include "gui/messages.h"
 #include "gui/string_utils.h"
 #include "gui/style_utils.h"
 #include "gui/styles.h"
